Created a Question::new() method
Created a new() method for Question objects to avoid having to call get_answer_prompt() every time a question is created. This is now called from within the new() method and only the neccissary data is stored in the Question object. It not longer stores the incorrect answers.
